move tests to top-level test directory
