+++ package.json
@@ -44,7 +44,7 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha './js/**/*.test.js' --compilers js:babel-core/register --recursive",
+    "test": "mocha --compilers js:babel-core/register --recursive",
     "test:watch": "npm test -- --watch",
     "start": "cross-env NODE_ENV=development node server.dev.js",
     "build": "npm run test && cross-env NODE_ENV=production webpack --config webpack.prod.config.js --progress --colors -p"
